Fixed migration in BitUtil from byte to int. Adjusted test coverage accordingly

+++ src/main/java/net/seblit/packeteer/BitUtil.java
@@ -20,7 +20,7 @@ public class BitUtil {
         for (int index : indices) {
             result |= (1 << index);
         }
-        return (byte) result;
+        return result;
     }
 
     /**
